--- a/cpp/arcticdb/codec/segment.hpp
+++ b/cpp/arcticdb/codec/segment.hpp
@@ -7,12 +7,13 @@
 
 #pragma once
 
+//#include <google/protobuf/io/zero_copy_stream_impl.h>  // lines 13-13
+//#include <codec/encoding_sizes.hpp>  // lines 12-12
+//#include <util/buffer_pool.hpp>  // lines 15-15
+
 #include "util/buffer.hpp"
 #include <arcticdb/storage/common.hpp>
-#include <codec/encoding_sizes.hpp>
-#include <google/protobuf/io/zero_copy_stream_impl.h>
 #include <google/protobuf/arena.h>
-#include <util/buffer_pool.hpp>
 #include <arcticdb/entity/field_collection.hpp>
 
 #include <iostream>