--- a/cpp/arcticdb/column_store/chunked_buffer.hpp
+++ b/cpp/arcticdb/column_store/chunked_buffer.hpp
@@ -15,6 +15,10 @@
 
 #include <cstdint>
 
+#ifndef DEBUG_BUILD
+#include <boost/container/small_vector.hpp>
+#endif
+
 namespace arcticdb {
 
 /*