--- a/cpp/arcticdb/entity/data_error.cpp
+++ b/cpp/arcticdb/entity/data_error.cpp
@@ -7,8 +7,6 @@
 
 #include <arcticdb/entity/data_error.hpp>
 
-#include <fmt/format.h>
-
 namespace arcticdb::entity {
 
 DataError::DataError(StreamId symbol,