--- a/cpp/arcticdb/entity/metrics.cpp
+++ b/cpp/arcticdb/entity/metrics.cpp
@@ -6,10 +6,11 @@
  */
 
 #include <arcticdb/entity/metrics.hpp>
+#include <prometheus/exposer.h>
+#include <prometheus/gateway.h>
 #include <arcticdb/log/log.hpp>
-#include <arcticdb/util/preconditions.hpp>
-#include <arcticdb/util/preprocess.hpp>
-#include <arcticdb/util/pb_util.hpp>
+#include <prometheus/registry.h>
+#include <prometheus/summary.h>
 
 #ifdef _WIN32
 #    include <Winsock.h> // for gethostname