--- a/cpp/arcticdb/entity/ref_key.hpp
+++ b/cpp/arcticdb/entity/ref_key.hpp
@@ -10,6 +10,7 @@
 #include <arcticdb/entity/types.hpp>
 #include <fmt/format.h>
 #include <arcticdb/entity/key.hpp>
+#include <arcticdb/util/hash.hpp>
 
 namespace arcticdb::entity {
     class RefKey {