--- a/cpp/arcticdb/entity/types.hpp
+++ b/cpp/arcticdb/entity/types.hpp
@@ -13,7 +13,6 @@
 #include <cstdint>
 #include <vector>
 #include <string>
-#include <type_traits>
 #include <variant>