--- a/cpp/arcticdb/log/log.cpp
+++ b/cpp/arcticdb/log/log.cpp
@@ -14,18 +14,14 @@
 #include <spdlog/sinks/basic_file_sink.h>
 #include <spdlog/sinks/rotating_file_sink.h>
 #include <spdlog/details/periodic_worker.h>
-#include <arcticdb/util/configs_map.hpp>
 #include <filesystem>
 
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/async.h>
 #include <spdlog/async_logger.h>
-#include <spdlog/sinks/stdout_sinks.h>
-
 
 #include <logger.pb.h>
 
-#include <memory>
 #include <mutex>
 #include <optional>