--- a/cpp/arcticdb/pipeline/index_segment_reader.hpp
+++ b/cpp/arcticdb/pipeline/index_segment_reader.hpp
@@ -8,13 +8,7 @@
 #pragma once
 
 #include <arcticdb/column_store/memory_segment.hpp>
-#include <arcticdb/entity/protobufs.hpp>
 #include <arcticdb/pipeline/frame_slice.hpp>
-#include <arcticdb/pipeline/index_fields.hpp>
-
-#include <boost/noncopyable.hpp>
-
-#include <cstdint>
 
 namespace arcticdb {
 class Store;
@@ -24,8 +18,10 @@ namespace arcticdb::pipelines {
 struct ReadQuery;
 }
 
-namespace arcticdb::pipelines::index {
 
+
+namespace arcticdb::pipelines::index {
+enum class Fields : uint32_t;
 struct IndexSegmentIterator;
 
 struct IndexSegmentReader {